+++ google/cloud/spanner/admin/internal/database_admin_stub.h
@@ -223,6 +223,13 @@ class DatabaseAdminStub {
       google::spanner::admin::database::v1::ListBackupSchedulesRequest const&
           request) = 0;
 
+  virtual StatusOr<google::spanner::admin::database::v1::
+                       InternalUpdateGraphOperationResponse>
+  InternalUpdateGraphOperation(
+      grpc::ClientContext& context, Options const& options,
+      google::spanner::admin::database::v1::
+          InternalUpdateGraphOperationRequest const& request) = 0;
+
   virtual StatusOr<google::longrunning::ListOperationsResponse> ListOperations(
       grpc::ClientContext& context, Options const& options,
       google::longrunning::ListOperationsRequest const& request) = 0;
@@ -441,6 +448,13 @@ class DefaultDatabaseAdminStub : public DatabaseAdminStub {
       grpc::ClientContext& context, Options const& options,
       google::spanner::admin::database::v1::ListBackupSchedulesRequest const&
           request) override;
+
+  StatusOr<google::spanner::admin::database::v1::
+               InternalUpdateGraphOperationResponse>
+  InternalUpdateGraphOperation(
+      grpc::ClientContext& context, Options const& options,
+      google::spanner::admin::database::v1::
+          InternalUpdateGraphOperationRequest const& request) override;
 
   StatusOr<google::longrunning::ListOperationsResponse> ListOperations(
       grpc::ClientContext& context, Options const& options,